Popups and artist events work (whoooo)
(Though real testing requires use of the console: `yayMap.set_artist('blah')`)

Interesting problems solved:

- Because the D3 overlay isn't a real Leaflet layer, when I tried to make it do a popup, the event bubbled up to the actual Leaflet layers and canceled the popup as soon as it had appeared.
- Because the example only sizes the SVG canvas to cover the geographic bounds, I had to add extra padding for the width of the circles.

Will it actually be better to use a second GeoJSON layer?  Perhaps.  If this makes it easier to animate the size changes, it'll be good.  Otherwise, I may go back to a normal layer because it's way easier, has zoom animations, and probably has better performance for lots of tour stops.